Renamed core API functions & type
The class NetBidi has been renamed to Bidi. This is because according to
Microsoft's guidelines a type should not have the same name as a
namespace. This means "using NetBidi" directives would actually be
functional.

BidiResolveString(Ex) has been renamed to ReorderString(Ex). Reorder is
more clear than Resolve because it more clearly communicates what the
functions do (They change the order (reorder) characters in a string).
The Bidi nature of the methods is clear in the renamed class, as a qualified call would be
(NetBidi.)Bidi.ReorderString(Ex)

This should make function invocations significantly more clear.