Major API Refactor - Using BidiString
- Renamed BidiStringData to BidiParagraphData

After a lot of thought I have decided to completely separate the logic
that's exposed to the user. There is the initial function that creates a
BidiString from which the user can just get the reordered string.

This allows me as the library creator embed information between function
calls and allows the user to add line information if/when they get it.

Although I must admit that I am not a fan of the current line API,
because sending the same lines seems to be a complete waste of memory in
my opinion. I am thinking of replacing it with an array of line lengths,
which is theoretically more efficient (One uint for line instead of
char).

I kind of need to look at the information soft-wrapping usually gives
users.